pyproject.toml
[project]
name = "pythonbankcards"
version = "0.1.0"
description = ""
authors = [
    {name = "Admin",email = "stiker1503@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest-cov = "^6.0.0"

